services/runner/requirements.txt
fastapi==0.128.1
uvicorn==0.40.0
httpx==0.27.0
pillow==10.4.0
boto3
boto3==1.34.162

